user_area/smbanaie/smbanaie/requirements.txt
pelican
markdown
pygments
pyyaml
ghp-import 
invoke 
pysvg 
requests
webassets
fab
google-api-python-client 
oauth2client
jdatetime